osf_demo_01/04_flask_app/flasklogin-tutorial/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Assets==2.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyMySQL==1.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
webassets==2.0
Werkzeug==2.3.4
WTForms==3.0.1